+++ lib/src/Pos/Block/Worker.hs
@@ -51,10 +51,11 @@ import qualified Pos.Lrc.DB as LrcDB (getLeadersForEpoch)
 import           Pos.Reporting (MetricMonitor (..), MetricMonitorState, noReportMonitor,
                                 recordValue, reportOrLogE)
 import           Pos.Slotting (currentTimeSlotting, getSlotStartEmpatically)
-import           Pos.Util (logWarningSWaitLinear, mconcatPair)
+import           Pos.Util (mconcatPair)
 import           Pos.Util.Chrono (OldestFirst (..))
 import           Pos.Util.JsonLog (jlCreatedBlock)
 import           Pos.Util.LogSafe (logDebugS, logInfoS, logWarningS)
+import           Pos.Util.TimeLimit (logWarningSWaitLinear)
 import           Pos.Util.Timer (Timer)
 import           Pos.Util.TimeWarp (CanJsonLog (..))
 import           Pos.WorkMode.Class (WorkMode)
