+++ lib/src/Pos/Lrc/Worker.hs
@@ -48,13 +48,13 @@ import           Pos.Ssc (MonadSscMem, noReportNoSecretsForEpoch1, sscCalculateS
 import           Pos.Ssc.Message (SscMessageConstraints)
 import           Pos.Update.DB (getCompetingBVStates)
 import           Pos.Update.Poll.Types (BlockVersionState (..))
-import           Pos.Util (logWarningWaitLinear, maybeThrow)
+import           Pos.Util (maybeThrow)
 import           Pos.Util.Chrono (NE, NewestFirst (..), toOldestFirst)
-
+import           Pos.Util.TimeLimit (logWarningWaitLinear)
 
 
 ----------------------------------------------------------------------------
--- Single short
+-- Single shot
 ----------------------------------------------------------------------------
 
 -- | 'LrcModeFull' contains all constraints necessary to launch LRC.
