+++ tools/cardano-sl-tools.cabal
@@ -97,6 +97,7 @@ executable cardano-genupdate
                      , ansi-wl-pprint
                      , bytestring
                      , cardano-sl
+                     , cardano-sl-util
                      , cryptonite
                      , neat-interpolation
                      , optparse-applicative
@@ -367,6 +368,7 @@ executable cardano-post-mortem
                        , cardano-sl
                        , cardano-sl-core
                        , cardano-sl-txp
+                       , cardano-sl-util
                        , Chart
                        , Chart-diagrams
                        , containers
@@ -429,12 +431,13 @@ executable cardano-blockchain-analyser
   if !flag(for-installer)
    build-depends:      ansi-wl-pprint
                      , base
+                     , cardano-sl
                      , cardano-sl-core
                      , cardano-sl-db
                      , cardano-sl-infra
                      , cardano-sl-lrc
                      , cardano-sl-txp
-                     , cardano-sl
+                     , cardano-sl-util
                      , data-default
                      , directory
                      , ether
