+++ util/Pos/Util.hs
@@ -1,0 +1,4 @@
+-- Pos.Util
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
+{-# OPTIONS_GHC -Wno-unused-imports   #-}
+{-# OPTIONS_GHC -Wno-dodgy-exports    #-}
