+++ util/Pos/Util/Util.hs
@@ -1,24 +1,13 @@
-{-# LANGUAGE CPP             #-}
-{-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE PolyKinds       #-}
-{-# LANGUAGE RankNTypes      #-}
+{-# LANGUAGE PolyKinds  #-}
+{-# LANGUAGE RankNTypes #-}
 
 module Pos.Util.Util
        (
-       -- * Something
-         Sign (..)
-       , maybeThrow
+       -- * Exceptions/errors
+         maybeThrow
        , eitherToFail
        , eitherToThrow
-       , getKeys
-       , sortWithMDesc
        , leftToPanic
-       , dumpSplices
-       , histogram
-       , median
-       , (<//>)
-       , divRoundUp
-       , sleep
 
        -- * Ether
        , ether
@@ -40,6 +29,24 @@ module Pos.Util.Util
        -- * Aeson
        , parseJSONWithRead
 
+       -- * NonEmpty
+       , neZipWith3
+       , neZipWith4
+       , spanSafe
+
+       -- * Misc
+       , mconcatPair
+       , microsecondsToUTC
+       , Sign (..)
+       , getKeys
+       , sortWithMDesc
+       , dumpSplices
+       , histogram
+       , median
+       , (<//>)
+       , divRoundUp
+       , sleep
+
        ) where
 
 import           Universum
@@ -51,20 +58,23 @@ import           Data.Aeson (FromJSON (..))
 import qualified Data.Aeson as A
 import qualified Data.Aeson.Types as A
 import           Data.HashSet (fromMap)
+import           Data.List (span, zipWith3, zipWith4)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map as M
+import           Data.Ratio ((%))
 import qualified Data.Semigroup as Smg
-import           Data.Time.Clock (NominalDiffTime)
+import           Data.Time.Clock (NominalDiffTime, UTCTime)
+import           Data.Time.Clock.POSIX (posixSecondsToUTCTime)
+import           Data.Time.Units (Microsecond, toMicroseconds)
 import qualified Ether
 import           Ether.Internal (HasLens (..))
 import qualified Language.Haskell.TH as TH
 import qualified Prelude
 
-----------------------------------------------------------------------------
--- Misc
-----------------------------------------------------------------------------
 
-data Sign = Plus | Minus
+----------------------------------------------------------------------------
+-- Exceptions/errors
+----------------------------------------------------------------------------
 
 maybeThrow :: (MonadThrow m, Exception e) => e -> Maybe a -> m a
 maybeThrow e = maybe (throwM e) pure
@@ -79,23 +89,15 @@ eitherToThrow
     => Either e a -> m a
 eitherToThrow = either throwM pure
 
--- | Create HashSet from HashMap's keys
-getKeys :: HashMap k v -> HashSet k
-getKeys = fromMap . void
-
--- | Use some monadic action to evaluate priority of value and sort a
--- list of values based on this priority. The order is descending
--- because I need it.
-sortWithMDesc :: (Monad m, Ord b) => (a -> m b) -> [a] -> m [a]
-sortWithMDesc f = fmap (map fst . sortWith (Down . snd)) . mapM f'
-  where
-    f' x = (x, ) <$> f x
-
 -- | Partial function which calls 'error' with meaningful message if
 -- given 'Left' and returns some value if given 'Right'.
 -- Intended usage is when you're sure that value must be right.
 leftToPanic :: Buildable a => Text -> Either a b -> b
 leftToPanic msgPrefix = either (error . mappend msgPrefix . pretty) identity
+
+----------------------------------------------------------------------------
+-- Ether
+----------------------------------------------------------------------------
 
 -- | Make a Reader or State computation work in an Ether transformer. Useful
 -- to make lenses work with Ether.
@@ -115,6 +117,10 @@ lensOf' = lensOf @a
 lensOfProxy :: forall proxy tag a b. HasLens tag a b => proxy tag -> Lens' a b
 lensOfProxy _ = lensOf @tag
 
+----------------------------------------------------------------------------
+-- PowerLift
+----------------------------------------------------------------------------
+
 class PowerLift m n where
     powerLift :: m a -> n a
 
@@ -123,6 +129,80 @@ instance {-# OVERLAPPING #-} PowerLift m m where
 
 instance (MonadTrans t, PowerLift m n, Monad n) => PowerLift m (t n) where
   powerLift = lift . powerLift @m @n
+
+----------------------------------------------------------------------------
+-- MinMax
+----------------------------------------------------------------------------
+
+newtype MinMax a = MinMax (Smg.Option (Smg.Min a, Smg.Max a))
+    deriving (Monoid)
+
+_MinMax :: Iso' (MinMax a) (Maybe (a, a))
+_MinMax = coerced
+
+mkMinMax :: a -> MinMax a
+mkMinMax a = _MinMax # Just (a, a)
+
+minMaxOf :: Getting (MinMax a) s a -> s -> Maybe (a, a)
+minMaxOf l = view _MinMax . foldMapOf l mkMinMax
+
+----------------------------------------------------------------------------
+-- Aeson
+----------------------------------------------------------------------------
+
+-- | Parse a value represented as a 'show'-ed string in JSON.
+parseJSONWithRead :: Read a => A.Value -> A.Parser a
+parseJSONWithRead =
+    either (fail . toString) pure . readEither @String <=<
+    parseJSON
+
+----------------------------------------------------------------------------
+-- NonEmpty
+----------------------------------------------------------------------------
+
+neZipWith3 :: (x -> y -> z -> q) -> NonEmpty x -> NonEmpty y -> NonEmpty z -> NonEmpty q
+neZipWith3 f (x :| xs) (y :| ys) (z :| zs) = f x y z :| zipWith3 f xs ys zs
+
+neZipWith4 ::
+       (x -> y -> i -> z -> q)
+    -> NonEmpty x
+    -> NonEmpty y
+    -> NonEmpty i
+    -> NonEmpty z
+    -> NonEmpty q
+neZipWith4 f (x :| xs) (y :| ys) (i :| is) (z :| zs) = f x y i z :| zipWith4 f xs ys is zs
+
+-- | Makes a span on the list, considering tail only. Predicate has
+-- list head as first argument. Used to take non-null prefix that
+-- depends on the first element.
+spanSafe :: (a -> a -> Bool) -> NonEmpty a -> (NonEmpty a, [a])
+spanSafe p (x:|xs) = let (a,b) = span (p x) xs in (x:|a,b)
+
+----------------------------------------------------------------------------
+-- Misc
+----------------------------------------------------------------------------
+
+-- | Specialized version of 'mconcat' (or 'Data.Foldable.fold')
+-- for restricting type to list of pairs.
+mconcatPair :: (Monoid a, Monoid b) => [(a, b)] -> (a, b)
+mconcatPair = mconcat
+
+microsecondsToUTC :: Microsecond -> UTCTime
+microsecondsToUTC = posixSecondsToUTCTime . fromRational . (% 1000000) . toMicroseconds
+
+data Sign = Plus | Minus
+
+-- | Create HashSet from HashMap's keys
+getKeys :: HashMap k v -> HashSet k
+getKeys = fromMap . void
+
+-- | Use some monadic action to evaluate priority of value and sort a
+-- list of values based on this priority. The order is descending
+-- because I need it.
+sortWithMDesc :: (Monad m, Ord b) => (a -> m b) -> [a] -> m [a]
+sortWithMDesc f = fmap (map fst . sortWith (Down . snd)) . mapM f'
+  where
+    f' x = (x, ) <$> f x
 
 -- | Concatenates two url parts using regular slash '/'.
 -- E.g. @"./dir/" <//> "/file" = "./dir/file"@.
@@ -177,27 +257,3 @@ median l = NE.sort l NE.!! middle
 -}
 sleep :: MonadIO m => NominalDiffTime -> m ()
 sleep n = liftIO (threadDelay (truncate (n * 10^(6::Int))))
-
--- MinMax
-
-newtype MinMax a = MinMax (Smg.Option (Smg.Min a, Smg.Max a))
-    deriving (Monoid)
-
-_MinMax :: Iso' (MinMax a) (Maybe (a, a))
-_MinMax = coerced
-
-mkMinMax :: a -> MinMax a
-mkMinMax a = _MinMax # Just (a, a)
-
-minMaxOf :: Getting (MinMax a) s a -> s -> Maybe (a, a)
-minMaxOf l = view _MinMax . foldMapOf l mkMinMax
-
-----------------------------------------------------------------------------
--- Aeson
-----------------------------------------------------------------------------
-
--- | Parse a value represented as a 'show'-ed string in JSON.
-parseJSONWithRead :: Read a => A.Value -> A.Parser a
-parseJSONWithRead =
-    either (fail . toString) pure . readEither @String <=<
-    parseJSON
